tinycms analytics: page views component
